Moves http related stuff into own module
