+++ docker.cabal
@@ -17,8 +17,8 @@ stability:           experimental
 library
   default-extensions:          DeriveGeneric, OverloadedStrings, ScopedTypeVariables, ExplicitForAll
   hs-source-dirs:      src
-  exposed-modules:     Docker, Docker.Types, Docker.Internal
-  -- other-modules:       Network.Docker.Internal
+  exposed-modules:     Docker, Docker.Types, Docker.Internal, Docker.Http
+  -- other-modules:       Docker.Internal
   build-depends:       base >= 4.7 && < 5
                      , lens
                      , aeson
